refactored plugin models to their own file
